requirements.txt
pyrogram
requests
tgcrypto
pymongo[srv]
python-dotenv
bs4
